requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
passlib
marshmallow
marshmallow-sqlalchemy
mysql-connector-python
bcrypt